requirements.txt
aiogram==2.25.1
openai==0.28
python-dotenv